feature-gate embedding,autotune,external-indexing jobs in daemon, so no additional dependencies will be compiled if the code is not used. Review all dependencies for lantern-cli in Cargo.toml and mark optional based on features that needs them
