+++ lantern_cli/Cargo.toml
@@ -12,11 +12,11 @@ path = "src/main.rs"
 [dependencies]
 clap = { version = "4.5.20", features = ["derive"] }
 anyhow = "1.0.91"
-postgres = "0.19.9"
+postgres = { version = "0.19.9", optional = true }
 rand = "0.8.5"
 linfa-clustering = { version = "0.7.0", features = ["ndarray-linalg"], optional = true }
 linfa = {version = "0.7.0", optional = true}
-ndarray = { version = "0.15.6", features = ["rayon"] }
+ndarray = { version = "0.15.6", features = ["rayon"], optional = true }
 rayon = { version="1.10.0", optional = true }
 md5 = {version="0.7.0", optional = true }
 serde = { version = "1.0", features = ["derive"] }
@@ -27,19 +27,18 @@ futures = "0.3.31"
 tokio = { version = "1.41.0", features = ["full"] }
 lazy_static = "1.5.0"
 itertools = "0.13.0"
-csv = "1.3.0"
-sysinfo = "0.32.0"
-tiktoken-rs = "0.6.0"
-url = "2.5"
-num_cpus = "1.16.0"
-ort = { version = "1.16.3", features = ["load-dynamic", "cuda", "openvino"] }
-tokenizers = { version = "0.20.1", features = ["default"] }
-image = { version = "0.25.4", features = ["jpeg", "png", "webp" ]}
-nvml-wrapper = "0.10.0"
-strum = { version = "0.26", features = ["derive"] }
-regex = "1.11.1"
-postgres-types = { version = "0.2.8", features = ["derive"] }
-usearch = { git = "https://github.com/Ngalstyan4/usearch.git", rev = "aa4f91d21230fd611b6c7741fa06be8c20acc9a9" }
+sysinfo = { version = "0.32.0", optional = true }
+tiktoken-rs = { version = "0.6.0", optional = true }
+url = { version = "2.5", optional = true }
+num_cpus = { version = "1.16.0", optional = true }
+ort = { version = "1.16.3", features = ["load-dynamic", "cuda", "openvino"], optional = true }
+tokenizers = { version = "0.20.1", features = ["default"], optional = true }
+image = { version = "0.25.4", features = ["jpeg", "png", "webp" ], optional = true }
+nvml-wrapper = { version = "0.10.0", optional = true }
+strum = { version = "0.26", features = ["derive"], optional = true }
+regex = { version = "1.11.1", optional = true }
+postgres-types = { version = "0.2.8", features = ["derive"], optional = true }
+usearch = { git = "https://github.com/Ngalstyan4/usearch.git", rev = "aa4f91d21230fd611b6c7741fa06be8c20acc9a9", optional = true }
 actix-web = { version = "4.9.0", optional = true }
 env_logger = { version = "0.11.5", optional = true }
 deadpool-postgres = { version = "0.14.0", optional = true }
@@ -53,18 +52,18 @@ bitvec = { version="1.0.1", optional=true }
 rustls = { version="0.23.16", optional=true }
 rustls-pemfile = { version="2.2.0", optional=true }
 glob = { version="0.3.1", optional=true }
-reqwest = { version = "0.12.9", default-features = false, features = ["json", "blocking", "rustls-tls"] }
+reqwest = { version = "0.12.9", default-features = false, features = ["json", "blocking", "rustls-tls"], optional = true }
 
 [features]
 default = ["cli", "daemon", "http-server", "autotune", "pq", "external-index", "external-index-server", "embeddings"]
 daemon = ["dep:tokio-postgres"]
-http-server = ["dep:deadpool-postgres", "dep:deadpool", "dep:bytes", "dep:utoipa", "dep:utoipa-swagger-ui", "dep:actix-web", "dep:tokio-postgres", "dep:env_logger", "dep:actix-web-httpauth"]
+http-server = ["dep:deadpool-postgres", "dep:deadpool", "dep:bytes", "dep:utoipa", "dep:utoipa-swagger-ui", "dep:actix-web", "dep:tokio-postgres", "dep:env_logger", "dep:actix-web-httpauth", "dep:regex"]
 autotune = []
-pq = ["dep:gcp_auth", "dep:linfa", "dep:linfa-clustering", "dep:md5", "dep:rayon"]
+pq = ["dep:gcp_auth", "dep:linfa", "dep:linfa-clustering", "dep:md5", "dep:rayon", "dep:reqwest", "dep:postgres", "dep:ndarray"]
 cli = []
-external-index = []
-external-index-server = ["dep:bitvec", "dep:rustls", "dep:rustls-pemfile", "dep:glob"]
-embeddings = ["dep:bytes"]
+external-index = ["dep:postgres-types", "dep:usearch", "dep:postgres"]
+external-index-server = ["dep:bitvec", "dep:rustls", "dep:rustls-pemfile", "dep:glob", "dep:usearch"]
+embeddings = ["dep:bytes", "dep:sysinfo", "dep:tiktoken-rs", "dep:url", "dep:num_cpus", "dep:ort", "dep:tokenizers", "dep:image", "dep:nvml-wrapper", "dep:strum", "dep:regex", "dep:reqwest", "dep:ndarray"]
 
 [lib]
 doctest = false
