+++ lantern_cli/src/cli.rs
@@ -1,11 +1,11 @@
-use super::daemon::cli::DaemonArgs;
-use super::embeddings::cli::{EmbeddingArgs, MeasureModelSpeedArgs, ShowModelsArgs};
-use super::external_index::cli::CreateIndexArgs;
-use super::http_server::cli::HttpServerArgs;
-use super::index_autotune::cli::IndexAutotuneArgs;
-use super::pq::cli::PQArgs;
 use clap::{Parser, Subcommand};
+use lantern_cli::daemon::cli::DaemonArgs;
+use lantern_cli::embeddings::cli::{EmbeddingArgs, MeasureModelSpeedArgs, ShowModelsArgs};
+use lantern_cli::external_index::cli::CreateIndexArgs;
 use lantern_cli::external_index::cli::IndexServerArgs;
+use lantern_cli::http_server::cli::HttpServerArgs;
+use lantern_cli::index_autotune::cli::IndexAutotuneArgs;
+use lantern_cli::pq::cli::PQArgs;
 
 #[derive(Subcommand, Debug)]
 pub enum Commands {
