+++ lantern_cli/src/daemon/types.rs
@@ -1,18 +1,8 @@
-use crate::embeddings::cli::{EmbeddingArgs, EmbeddingJobType, Runtime};
-use crate::embeddings::core::utils::get_clean_model_name;
-use crate::external_index::cli::CreateIndexArgs;
-use crate::index_autotune::cli::IndexAutotuneArgs;
-use crate::logger::Logger;
-use crate::types::{AnyhowVoidResult, ProgressCbFn};
-use crate::utils::{get_common_embedding_ignore_filters, quote_literal};
-use itertools::Itertools;
 use std::collections::HashMap;
-use std::sync::Arc;
 use tokio::sync::{
     mpsc::{Sender, UnboundedSender},
     Mutex, RwLock,
 };
-use tokio_postgres::Row;
 use tokio_util::sync::CancellationToken;
 
 #[derive(Clone)]
@@ -54,170 +44,6 @@ impl TargetDB {
     }
 }
 
-#[derive(Debug, Clone)]
-pub struct EmbeddingJob {
-    pub id: i32,
-    pub is_init: bool,
-    pub db_uri: String,
-    pub schema: String,
-    pub table: String,
-    pub column: String,
-    pub pk: String,
-    pub filter: Option<String>,
-    pub label: Option<String>,
-    pub job_type: EmbeddingJobType,
-    pub column_type: String,
-    pub out_column: String,
-    pub model: String,
-    pub runtime_params: String,
-    pub runtime: Runtime,
-    pub batch_size: Option<usize>,
-    pub row_ids: Option<Vec<String>>,
-}
-
-impl EmbeddingJob {
-    pub fn new(row: Row, data_path: &str, db_uri: &str) -> Result<EmbeddingJob, anyhow::Error> {
-        let runtime = Runtime::try_from(row.get::<&str, Option<&str>>("runtime").unwrap_or("ort"))?;
-        let runtime_params = if runtime == Runtime::Ort {
-            format!(r#"{{ "data_path": "{data_path}" }}"#)
-        } else {
-            row.get::<&str, Option<String>>("runtime_params")
-                .unwrap_or("{}".to_owned())
-        };
-
-        let batch_size = if let Some(batch_size) = row.get::<&str, Option<i32>>("batch_size") {
-            Some(batch_size as usize)
-        } else {
-            None
-        };
-
-        Ok(Self {
-            id: row.get::<&str, i32>("id"),
-            pk: row.get::<&str, String>("pk"),
-            label: row.get::<&str, Option<String>>("label"),
-            db_uri: db_uri.to_owned(),
-            schema: row.get::<&str, String>("schema"),
-            table: row.get::<&str, String>("table"),
-            column: row.get::<&str, String>("column"),
-            out_column: row.get::<&str, String>("dst_column"),
-            model: get_clean_model_name(row.get::<&str, &str>("model"), runtime),
-            runtime,
-            runtime_params,
-            filter: None,
-            row_ids: None,
-            is_init: true,
-            batch_size,
-            job_type: EmbeddingJobType::try_from(
-                row.get::<&str, Option<&str>>("job_type")
-                    .unwrap_or("embedding"),
-            )?,
-            column_type: row
-                .get::<&str, Option<String>>("column_type")
-                .unwrap_or("REAL[]".to_owned()),
-        })
-    }
-
-    pub fn set_filter(&mut self, filter: &str) {
-        self.filter = Some(filter.to_owned());
-    }
-
-    pub fn set_is_init(&mut self, is_init: bool) {
-        self.is_init = is_init;
-    }
-
-    pub fn set_row_ids(&mut self, row_ids: Vec<String>) {
-        self.row_ids = Some(row_ids);
-    }
-
-    #[allow(dead_code)]
-    pub fn set_ctid_filter(&mut self, row_ids: &Vec<String>) {
-        let row_ctids_str = row_ids
-            .iter()
-            .map(|r| {
-                format!(
-                    "currtid2('{table_name}','{r}'::tid)",
-                    table_name = &self.table
-                )
-            })
-            .join(",");
-        self.set_filter(&format!("ctid IN ({row_ctids_str})"));
-    }
-
-    pub fn set_id_filter(&mut self, row_ids: &Vec<String>) {
-        let row_ctids_str = row_ids.iter().map(|s| quote_literal(s)).join(",");
-        self.set_filter(&format!(
-            "id IN ({row_ctids_str}) AND {common_filter}",
-            common_filter = get_common_embedding_ignore_filters(&self.column)
-        ));
-    }
-}
-
-#[derive(Debug)]
-pub struct AutotuneJob {
-    pub id: i32,
-    pub is_init: bool,
-    pub db_uri: String,
-    pub schema: String,
-    pub table: String,
-    pub column: String,
-    pub metric_kind: String,
-    pub model_name: Option<String>,
-    pub recall: f64,
-    pub k: u16,
-    pub sample_size: usize,
-    pub create_index: bool,
-}
-
-impl AutotuneJob {
-    pub fn new(row: Row, db_uri: &str) -> AutotuneJob {
-        Self {
-            id: row.get::<&str, i32>("id"),
-            db_uri: db_uri.to_owned(),
-            schema: row.get::<&str, String>("schema"),
-            table: row.get::<&str, String>("table"),
-            column: row.get::<&str, String>("column"),
-            metric_kind: row.get::<&str, String>("metric_kind"),
-            model_name: row.get::<&str, Option<String>>("model"),
-            recall: row.get::<&str, f64>("target_recall"),
-            k: row.get::<&str, i32>("k") as u16,
-            sample_size: row.get::<&str, i32>("sample_size") as usize,
-            create_index: row.get::<&str, bool>("create_index"),
-            is_init: true,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct ExternalIndexJob {
-    pub id: i32,
-    pub db_uri: String,
-    pub schema: String,
-    pub table: String,
-    pub column: String,
-    pub operator_class: String,
-    pub index_name: Option<String>,
-    pub ef: usize,
-    pub efc: usize,
-    pub m: usize,
-}
-
-impl ExternalIndexJob {
-    pub fn new(row: Row, db_uri: &str) -> ExternalIndexJob {
-        Self {
-            id: row.get::<&str, i32>("id"),
-            db_uri: db_uri.to_owned(),
-            schema: row.get::<&str, String>("schema"),
-            table: row.get::<&str, String>("table"),
-            column: row.get::<&str, String>("column"),
-            operator_class: row.get::<&str, String>("operator"),
-            index_name: row.get::<&str, Option<String>>("index"),
-            ef: row.get::<&str, i32>("ef") as usize,
-            efc: row.get::<&str, i32>("efc") as usize,
-            m: row.get::<&str, i32>("m") as usize,
-        }
-    }
-}
-
 #[derive(Debug)]
 pub struct JobInsertNotification {
     pub id: i32,
@@ -249,32 +75,43 @@ pub enum ClientJobSignal {
     Restart,
 }
 
+#[cfg(feature = "embeddings")]
 pub type EmbeddingProcessorArgs = (
-    EmbeddingArgs,
+    crate::embeddings::cli::EmbeddingArgs,
     Sender<Result<(usize, usize), anyhow::Error>>,
-    Logger,
+    crate::logger::Logger,
 );
+#[cfg(not(feature = "embeddings"))]
+pub type EmbeddingProcessorArgs = ();
 
+#[cfg(feature = "autotune")]
 pub type AutotuneProcessorArgs = (
-    IndexAutotuneArgs,
-    Sender<AnyhowVoidResult>,
+    crate::index_autotune::cli::IndexAutotuneArgs,
+    Sender<crate::types::AnyhowVoidResult>,
     JobTaskEventTx,
-    Option<ProgressCbFn>,
-    Arc<std::sync::RwLock<bool>>,
-    Logger,
+    Option<crate::types::ProgressCbFn>,
+    std::sync::Arc<std::sync::RwLock<bool>>,
+    crate::logger::Logger,
 );
+#[cfg(not(feature = "autotune"))]
+pub type AutotuneProcessorArgs = ();
 
+#[cfg(feature = "external-index")]
 pub type ExternalIndexProcessorArgs = (
-    CreateIndexArgs,
-    Sender<AnyhowVoidResult>,
+    crate::external_index::cli::CreateIndexArgs,
+    Sender<crate::types::AnyhowVoidResult>,
     JobTaskEventTx,
-    Option<ProgressCbFn>,
-    Arc<std::sync::RwLock<bool>>,
-    Logger,
+    Option<crate::types::ProgressCbFn>,
+    std::sync::Arc<std::sync::RwLock<bool>>,
+    crate::logger::Logger,
 );
+#[cfg(not(feature = "external-index"))]
+pub type ExternalIndexProcessorArgs = ();
 
 pub enum JobType {
     Embeddings(Sender<EmbeddingProcessorArgs>),
+    #[allow(dead_code)]
     ExternalIndex(Sender<ExternalIndexProcessorArgs>),
+    #[allow(dead_code)]
     Autotune(Sender<AutotuneProcessorArgs>),
 }
