+++ lantern_cli/src/embeddings/measure_speed.rs
@@ -5,7 +5,7 @@ use super::{
     core::{EmbeddingRuntime, Runtime},
 };
 use crate::logger::{LogLevel, Logger};
-use postgres::NoTls;
+use tokio_postgres::NoTls;
 use tokio_util::sync::CancellationToken;
 
 use super::cli::MeasureModelSpeedArgs;
